requirements.txt
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
numpy>=1.24.0
scikit-learn>=1.2.0
redis>=5.0.0
onnxruntime>=1.14.0
pytorch-lightning>=2.0.0
pennylane>=0.31.0
qiskit>=0.43.0
tvm>=0.11.0
deap>=1.3.1
pandas>=2.0.0
scipy>=1.10.0
networkx>=3.1.0
Flask
datasets
Pillow
pytest
pytest-flask
